Split mirror snapshot inspection ownership
