+++ deploy/docker/Cargo.lock.hosted-provider
@@ -2744,6 +2744,7 @@ version = "0.1.0-beta.90"
 dependencies = [
  "async-trait",
  "axum",
+ "base64",
  "chrono",
  "fs2",
  "mdbase",
