+++ deploy/docker/mdbase-rs-revision
@@ -1,1 +1,1 @@
-36fdb863002804f11955f93921a7b768d8de43ae
+a26d5e6f0ffce42e5e83287025e9068adfdc94c6
